refactor: extract ssl-key-storage lifecycle into focused module
Split src/ssl-bump.ts into two modules:

- src/ssl-key-storage.ts (~140 lines): key-storage lifecycle
  - mountSslTmpfs / unmountSslTmpfs (tmpfs management)
  - chownRecursive (directory ownership)
  - secureWipeFile / cleanupSslKeyMaterial (secure cleanup)

- src/ssl-bump.ts (~100 lines, reduced from 304): CA generation & DB init
  - generateSessionCa (now delegates tmpfs mount)
  - initSslDb
  - isOpenSslAvailable
  - CaFiles interface

Existing callers continue to work via re-exports from ssl-bump.ts.

Closes #5633
